moved Search to separate component
